requirements.txt
PyYAML~=6.0.1
inflection~=0.5.1
decohints~=1.0.9
pydantic~=2.6.1